helicoptero completo

+++ src/main/java/ar/edu/unahur/obj2/excepciones/modos/Modo.java
@@ -1,9 +1,9 @@
 package ar.edu.unahur.obj2.excepciones.modos;
 
 public interface Modo {
-    double getConsumoPorLitro();
+    Double getConsumoPorLitro();
 
-    double getVelocidadMax();
+    Integer getVelocidadMax();
 
     String getNombre();
 }
