+++ src/main/java/ar/edu/unahur/obj2/excepciones/modos/ModoAgresivo.java
@@ -2,13 +2,13 @@ package ar.edu.unahur.obj2.excepciones.modos;
 
 public class ModoAgresivo implements Modo{
     @Override
-    public double getConsumoPorLitro(){
+    public Double getConsumoPorLitro(){
         return 6.0;
     }
 
     @Override
-    public double getVelocidadMax(){
-        return 280.0;
+    public Integer getVelocidadMax(){
+        return 280;
     }
 
     @Override
