+++ src/main/java/ar/edu/unahur/obj2/excepciones/modos/ModoNormal.java
@@ -2,13 +2,13 @@ package ar.edu.unahur.obj2.excepciones.modos;
 
 public class ModoNormal implements Modo{
     @Override
-    public double getConsumoPorLitro(){
+    public Double getConsumoPorLitro(){
         return 9.0;
     }
 
     @Override
-    public double getVelocidadMax(){
-        return 220.0;
+    public Integer getVelocidadMax(){
+        return 220;
     }
 
     @Override
